Improve treatment of presses
Press length=1 is unrealistic. Discard such presses. Similarly, if two
OC or two OT presses are only a few lines apart, then it is probably a
double-press indicating a long vehicle. We don't do this anymore, but
early files have such presses.

If a list of candidate OT intervals is empty when we're sure it's an
OT event, don't error out but simply record an empty OT interval. Same
for the OC case.