Fix split-and-strip approach to parts in partitions
When a part returned by Louvain (edges are based on distances, not
time; thus 1 part can contain two overtakes, if they happen to be
roughly at the same lat. dist.), split it to subparts around time
gaps, then discard subparts which are too small and cannot be
overtakes. Finally, take the the leftmost (closes to press) subpart as
the event.