Revert strip-and-split treatment of the parts
Some outlier values in a detected event could be excluded as they are
slightly off. However, this is a scaled down problem analogous to the
one we started out with: select a coherent sub-community from a given
set of values. This secondary problem is too small for the modularity
approach though. And other more primitive methods are going to have
unwanted side-effects. Better to leave this alone for now and accept
slightly exaggerated values here and there.